agentHost: scope todo telemetry to Copilot CLI
Move todo SQL classification and emission into the Copilot provider path so other harnesses do not participate. Keep generic tool-call correlation in the shared tracker.
