+++ src/vs/platform/agentHost/node/agentHostToolCallTracker.ts
@@ -6,7 +6,6 @@
 import { disposableTimeout } from '../../../base/common/async.js';
 import { Disposable, DisposableMap } from '../../../base/common/lifecycle.js';
 import { StopWatch } from '../../../base/common/stopwatch.js';
-import type { TodoStoreOperation, TodoStoreTarget } from '../../telemetry/common/todoStoreTelemetry.js';
 import type { SessionToolAuthenticationRequest, SessionToolClientExecutionRequest, SessionToolConfirmationRequest } from '../common/state/protocol/state.js';
 import { ToolCallContributorKind, type ToolCallContributor, type ToolCallResult } from '../common/state/sessionState.js';
 import type { AgentHostModelTelemetryKind, AgentHostTelemetryReporter, IAgentHostToolInvokedReport } from './agentHostTelemetryReporter.js';
@@ -76,7 +75,6 @@ interface IToolCallTiming {
 	model: string | undefined;
 	modelTelemetryKind: AgentHostModelTelemetryKind | undefined;
 	modelResolvedFromUsage: boolean;
-	todoStoreOperation: ITodoStoreOperationData | undefined;
 }
 
 interface IStalledToolCall {
@@ -123,7 +121,6 @@ export class AgentHostToolCallTracker extends Disposable {
 			model: resolvedModel?.model ?? model,
 			modelTelemetryKind: resolvedModel?.modelTelemetryKind ?? modelTelemetryKind,
 			modelResolvedFromUsage: resolvedModel !== undefined,
-			todoStoreOperation: undefined,
 		});
 	}
 
@@ -164,16 +161,6 @@ export class AgentHostToolCallTracker extends Disposable {
 		}
 	}
 
-	toolCallReady(session: string, toolCallId: string, toolInput: string | undefined): void {
-		if (toolInput === undefined) {
-			return;
-		}
-		const timing = this._toolCalls.get(this._key(session, toolCallId));
-		if (timing?.toolSourceKind === 'agentHost') {
-			timing.todoStoreOperation = getTodoStoreOperationData(timing.toolId, toolInput);
-		}
-	}
-
 	toolCallCompleted(session: string, toolCallId: string, result: ToolCallResult): void {
 		const key = this._key(session, toolCallId);
 		const timing = this._toolCalls.get(key);
@@ -209,16 +196,6 @@ export class AgentHostToolCallTracker extends Disposable {
 			pending.push(report);
 			this._pendingToolReports.set(turnKey, pending);
 		}
-		const todoStoreOperation = result.success ? timing.todoStoreOperation : undefined;
-		if (todoStoreOperation) {
-			this._reporter.todoStoreOperation({
-				provider: timing.provider,
-				session: timing.session,
-				toolCallId,
-				...todoStoreOperation,
-			});
-		}
-
 		const stalled = this._stalledToolCalls.get(key);
 		if (stalled) {
 			this._stalledToolCalls.delete(key);
@@ -313,211 +290,4 @@ export class AgentHostToolCallTracker extends Disposable {
 	private _turnKey(session: string, turnId: string): string {
 		return `${session}\0${turnId}`;
 	}
-}
-
-interface ITodoStoreOperationData {
-	readonly operation: TodoStoreOperation;
-	readonly target: TodoStoreTarget;
-}
-
-interface ISqlToken {
-	readonly value: string;
-	readonly kind: 'identifier' | 'punctuation';
-}
-
-export function getTodoStoreOperationData(toolId: string, toolInput: string | undefined): ITodoStoreOperationData | undefined {
-	if (toolId !== 'sql') {
-		return undefined;
-	}
-
-	const query = parseToolInput(toolInput)?.query;
-	if (typeof query !== 'string') {
-		return undefined;
-	}
-
-	const tokens = tokenizeSql(query);
-	const readTargets = new Set<TodoStoreTarget>();
-	const writeTargets = new Set<TodoStoreTarget>();
-	const deleteFromIndexes = new Set<number>();
-
-	for (let i = 0; i < tokens.length; i++) {
-		switch (tokens[i].value) {
-			case 'insert':
-			case 'replace': {
-				const intoIndex = findToken(tokens, i + 1, 'into', ['or', 'rollback', 'abort', 'replace', 'fail', 'ignore']);
-				addTodoStoreTarget(writeTargets, readTableIdentifier(tokens, intoIndex + 1));
-				break;
-			}
-			case 'update':
-				addTodoStoreTarget(writeTargets, readTableIdentifier(tokens, i + 1, ['or', 'rollback', 'abort', 'replace', 'fail', 'ignore']));
-				break;
-			case 'delete': {
-				const fromIndex = findToken(tokens, i + 1, 'from');
-				deleteFromIndexes.add(fromIndex);
-				addTodoStoreTarget(writeTargets, readTableIdentifier(tokens, fromIndex + 1));
-				break;
-			}
-			case 'create':
-			case 'drop':
-			case 'alter': {
-				const tableIndex = findToken(tokens, i + 1, 'table', ['temp', 'temporary']);
-				addTodoStoreTarget(writeTargets, readTableIdentifier(tokens, tableIndex + 1, ['if', 'not', 'exists']));
-				break;
-			}
-		}
-	}
-
-	for (let i = 0; i < tokens.length; i++) {
-		if (tokens[i].value === 'from' && !deleteFromIndexes.has(i)) {
-			addFromClauseTargets(tokens, i + 1, readTargets);
-		} else if (tokens[i].value === 'join') {
-			addTodoStoreTarget(readTargets, readTableIdentifier(tokens, i + 1));
-		}
-	}
-
-	if (readTargets.size === 0 && writeTargets.size === 0) {
-		return undefined;
-	}
-
-	const referencesTodos = readTargets.has('todos') || writeTargets.has('todos');
-	const referencesTodoDeps = readTargets.has('todo_deps') || writeTargets.has('todo_deps');
-	return {
-		operation: readTargets.size > 0 && writeTargets.size > 0 ? 'mixed' : writeTargets.size > 0 ? 'write' : 'read',
-		target: referencesTodos && referencesTodoDeps ? 'both' : referencesTodoDeps ? 'todo_deps' : 'todos',
-	};
-}
-
-function tokenizeSql(query: string): ISqlToken[] {
-	const tokens: ISqlToken[] = [];
-	for (let i = 0; i < query.length;) {
-		const char = query[i];
-		const next = query[i + 1];
-		if (/\s/.test(char)) {
-			i++;
-		} else if (char === '-' && next === '-') {
-			i = skipUntil(query, i + 2, '\n');
-		} else if (char === '/' && next === '*') {
-			i = skipUntil(query, i + 2, '*/');
-		} else if (char === '\'') {
-			i = skipQuoted(query, i + 1, '\'', '\'');
-		} else if (char === '"' || char === '`') {
-			const end = skipQuoted(query, i + 1, char, char);
-			tokens.push({ value: query.slice(i + 1, end - 1).replaceAll(char + char, char).toLowerCase(), kind: 'identifier' });
-			i = end;
-		} else if (char === '[') {
-			const end = skipQuoted(query, i + 1, ']', ']');
-			tokens.push({ value: query.slice(i + 1, end - 1).replaceAll(']]', ']').toLowerCase(), kind: 'identifier' });
-			i = end;
-		} else if (/[a-z_$]/i.test(char)) {
-			let end = i + 1;
-			while (end < query.length && /[\w$]/.test(query[end])) {
-				end++;
-			}
-			tokens.push({ value: query.slice(i, end).toLowerCase(), kind: 'identifier' });
-			i = end;
-		} else {
-			if (char === '.' || char === ',' || char === '(' || char === ')' || char === ';') {
-				tokens.push({ value: char, kind: 'punctuation' });
-			}
-			i++;
-		}
-	}
-	return tokens;
-}
-
-function skipUntil(query: string, start: number, terminator: string): number {
-	const index = query.indexOf(terminator, start);
-	return index === -1 ? query.length : index + terminator.length;
-}
-
-function skipQuoted(query: string, start: number, terminator: string, escape: string): number {
-	for (let i = start; i < query.length; i++) {
-		if (query[i] !== terminator) {
-			continue;
-		}
-		if (query[i + 1] === escape) {
-			i++;
-		} else {
-			return i + 1;
-		}
-	}
-	return query.length;
-}
-
-function findToken(tokens: readonly ISqlToken[], start: number, value: string, skippedValues: readonly string[] = []): number {
-	for (let i = start; i < tokens.length && tokens[i].value !== ';'; i++) {
-		if (tokens[i].value === value) {
-			return i;
-		}
-		if (!skippedValues.includes(tokens[i].value)) {
-			break;
-		}
-	}
-	return -1;
-}
-
-function readTableIdentifier(tokens: readonly ISqlToken[], start: number, skippedValues: readonly string[] = []): string | undefined {
-	let index = start;
-	while (index < tokens.length && skippedValues.includes(tokens[index].value)) {
-		index++;
-	}
-	if (tokens[index]?.kind !== 'identifier') {
-		return undefined;
-	}
-
-	let table = tokens[index].value;
-	while (tokens[index + 1]?.value === '.' && tokens[index + 2]?.kind === 'identifier') {
-		table = tokens[index + 2].value;
-		index += 2;
-	}
-	return table;
-}
-
-function addFromClauseTargets(tokens: readonly ISqlToken[], start: number, targets: Set<TodoStoreTarget>): void {
-	const terminators = new Set(['where', 'group', 'order', 'having', 'limit', 'union', 'intersect', 'except', 'returning', 'set', 'values', ';']);
-	let expectsTable = true;
-	let depth = 0;
-	for (let i = start; i < tokens.length; i++) {
-		const value = tokens[i].value;
-		if (value === '(') {
-			if (depth === 0 && expectsTable) {
-				expectsTable = false;
-			}
-			depth++;
-		} else if (value === ')') {
-			if (depth === 0) {
-				return;
-			}
-			depth--;
-		} else if (depth === 0 && terminators.has(value)) {
-			return;
-		} else if (depth === 0 && (value === ',' || value === 'join')) {
-			expectsTable = true;
-		} else if (depth === 0 && expectsTable && tokens[i].kind === 'identifier') {
-			addTodoStoreTarget(targets, readTableIdentifier(tokens, i));
-			expectsTable = false;
-		}
-	}
-}
-
-function addTodoStoreTarget(targets: Set<TodoStoreTarget>, identifier: string | undefined): void {
-	if (identifier === 'todos' || identifier === 'todo_deps') {
-		targets.add(identifier);
-	}
-}
-
-function parseToolInput(toolInput: string | undefined): Record<string, unknown> | undefined {
-	if (toolInput === undefined) {
-		return undefined;
-	}
-	try {
-		const parsed: unknown = JSON.parse(toolInput);
-		return isRecord(parsed) ? parsed : undefined;
-	} catch {
-		return undefined;
-	}
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-	return value !== null && typeof value === 'object' && !Array.isArray(value);
 }
