+++ src/vs/platform/agentHost/node/agentSideEffects.ts
@@ -809,10 +809,6 @@ export class AgentSideEffects extends Disposable {
 			this._toolCallTracker.updateTurnModel(sessionKey, action.turnId, modelContext.model, modelContext.modelTelemetryKind);
 		}
 
-		if (action.type === ActionType.ChatToolCallReady) {
-			this._toolCallTracker.toolCallReady(sessionKey, action.toolCallId, getInlineToolInput(action.toolInput));
-		}
-
 		const sessionUri = isAhpChatChannel(sessionKey) ? parseRequiredSessionUriFromChatUri(sessionKey) : sessionKey;
 
 		// Stamp the subagent chat URI onto the tool call as soon as toolKind
