+++ src/vs/platform/agentHost/node/copilot/copilotAgentSession.ts
@@ -77,6 +77,7 @@ import { McpAuthRequiredReason, McpServerStatus, type McpAuthRequirement, type M
 import type { ErrorInfo, ProtectedResourceMetadata } from '../../common/state/protocol/common/state.js';
 import { CopilotSlashCommandProvider } from './copilotSlashCommandProvider.js';
 import { createCopilotFailureCorrelation, reportCopilotModelCallFailure, reportCopilotSdkSessionError } from './copilotFailureTelemetry.js';
+import { reportCopilotTodoStoreOperation } from './copilotTodoStoreTelemetry.js';
 
 /**
  * The full set of agent modes the Copilot SDK accepts. AHP now exposes the
@@ -4042,6 +4043,12 @@ export class CopilotAgentSession extends Disposable {
 			if (!parentToolCallId && e.agentId) {
 				this._logService.warn(`[Copilot:${this.sessionId}] Dropping tool.execution_complete for unknown subagent agentId=${e.agentId}`);
 				return;
+			}
+			if (e.data.success && tracked.contributor === undefined) {
+				const telemetrySession = parentToolCallId
+					? URI.parse(buildSubagentSessionUri(this._storageUri.toString(), parentToolCallId))
+					: this.sessionUri;
+				reportCopilotTodoStoreOperation(this._telemetryService, telemetrySession, e.data.toolCallId, tracked.toolName, tracked.parameters);
 			}
 			this._logService.info(`[Copilot:${sessionId}] Tool completed: ${e.data.toolCallId}`);
 			this._reportToolApprovalIfNoPermission(e.data.toolCallId);
